+++ relay/adapter-express/tests/index.spec.ts
@@ -135,8 +135,13 @@ describe('RelayAdapterExpress', () => {
 
     const storage = new MemoryStorage();
     const authenticator = new Authenticator({ wallet: app, appInfo, chainInfo });
-    // eslint-disable-next-line no-console
-    const logger = { info: console.info, error: console.error, warn: console.warn, debug: noop };
+    const logger = {
+      // eslint-disable-next-line no-console
+      info: process.env.CI ? noop : console.info,
+      error: process.env.CI ? noop : console.error,
+      warn: process.env.CI ? noop : console.warn,
+      debug: noop,
+    };
     const handlers = createHandlers({ storage, authenticator, logger });
 
     handlers.wsServer.attach(server.http);